Small changes.
Add a Makefile to the root directory to simplify common tasks.

Update the standard version.

Update copyright so that there are no spaces in short license names.

Include man pages in the seec binary package.

Update the README.md to describe the submodules and building using the newly added Makefile.